Switched out useCustomMonaco for MonacoProvider that wraps the App. monaco is loaded but nothing is really initialized till the editors are loaded. Don't need custom code for prettier formatting (some more fixes coming soon), or theming, all built in), attempted typing, need to debug it
`useFile` and `useEditor` pick up monaco from context if they are not input. Also changed the file paths to .ts and .tsx so that the typscript worker can interpret them correctly. Might not be the desired behaviour but I was doing it to test the typings.

the onLoad bug will be fixed in a patch version, the current changes dont include it